requirements.txt
apache-beam==2.29.0
avro-python3==1.9.2.1
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.4.1
firebase-admin==5.0.0
firestore==0.0.8
future==0.18.2
google-api-core==1.27.0
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
hdfs==2.6.0
httplib2==0.17.4
idna==2.10
iso8601==0.1.14
msgpack==1.0.2
numpy==1.20.3
oauth2client==4.1.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
